Compilation fixes.

+++ sql/storage/bat/bat_storage.c
@@ -2929,7 +2929,7 @@ claim_segment(sql_trans *tr, sql_table *t, storage *s, size_t cnt)
 	if ((!inTransaction(tr, t) && !in_transaction && isGlobal(t)) || (!isNew(t) && isLocalTemp(t))) {
 		trans_add(tr, &t->base, s, &tc_gc_del, &commit_update_del, isLocalTemp(t)?NULL:&log_update_del);
 		if (!isLocalTemp(t))
-			tr->logchanges += cnt;
+			tr->logchanges += (int) cnt;
 	}
 	return slot;
 }
